restructure package for common module
