+++ src/main/java/com/qa/api_common/playground/Playground.java
@@ -1,0 +1,24 @@
+
+/*
+ * @Playground.java@
+ * Created on 08-Sep-2023
+ *
+ * Copyright (c) 2023 Imspaliwal.
+ * All Rights Reserved.
+ *
+ * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF Imspaliwal
+ * The copyright notice above does not evidence any
+ * actual or intended publication of such source code.
+ *
+ */
+package com.qa.api_common.playground;
+
+public class Playground {
+
+    public static void main(String[] args) {
+
+        System.out.println(System.getProperty("browser"));
+
+    }
+
+}
