+++ src/main/java/com/qa/api_common/screenshot/TakeScreenshot.java
@@ -1,0 +1,20 @@
+
+/*
+ * @TakeScreenshot.java@
+ * Created on 24-Sep-2023
+ *
+ * Copyright (c) 2023 Imspaliwal.
+ * All Rights Reserved.
+ *
+ * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF Imspaliwal
+ * The copyright notice above does not evidence any
+ * actual or intended publication of such source code.
+ *
+ */
+package com.qa.api_common.screenshot;
+
+public enum TakeScreenshot {
+
+    True, False;
+
+}
