+++ src/main/java/com/qa/api_common/testng/TestListener.java
@@ -1,0 +1,22 @@
+
+/*
+ * @TestListener.java@
+ * Created on 24-Sep-2023
+ *
+ * Copyright (c) 2023 Imspaliwal.
+ * All Rights Reserved.
+ *
+ * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF Imspaliwal
+ * The copyright notice above does not evidence any
+ * actual or intended publication of such source code.
+ *
+ */
+package com.qa.api_common.testng;
+
+public class TestListener {
+
+    public static void log(String msg) {
+        // System.out.println(msg);
+    }
+
+}
