--- src/main/java/com/qa/orangehrm/Playground.java
@@ -1,24 +1,0 @@
-
-/*
- * @Playground.java@
- * Created on 08-Sep-2023
- *
- * Copyright (c) 2023 Imspaliwal.
- * All Rights Reserved.
- *
- * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF Imspaliwal
- * The copyright notice above does not evidence any
- * actual or intended publication of such source code.
- *
- */
-package com.qa.orangehrm;
-
-public class Playground {
-
-    public static void main(String[] args) {
-
-        System.out.println(System.getProperty("browser"));
-
-    }
-
-}
