--- src/main/java/com/qa/orangehrm/screenshot/TakeScreenshot.java
@@ -1,20 +1,0 @@
-
-/*
- * @TakeScreenshot.java@
- * Created on 24-Sep-2023
- *
- * Copyright (c) 2023 Imspaliwal.
- * All Rights Reserved.
- *
- * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF Imspaliwal
- * The copyright notice above does not evidence any
- * actual or intended publication of such source code.
- *
- */
-package com.qa.orangehrm.screenshot;
-
-public enum TakeScreenshot {
-
-    True, False;
-
-}
