--- src/main/java/com/qa/orangehrm/testng/TestListener.java
@@ -1,22 +1,0 @@
-
-/*
- * @TestListener.java@
- * Created on 24-Sep-2023
- *
- * Copyright (c) 2023 Imspaliwal.
- * All Rights Reserved.
- *
- * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF Imspaliwal
- * The copyright notice above does not evidence any
- * actual or intended publication of such source code.
- *
- */
-package com.qa.orangehrm.testng;
-
-public class TestListener {
-
-    public static void log(String msg) {
-        // System.out.println(msg);
-    }
-
-}
