+++ src/main/java/com/qa/orangehrm/util/AppUtil.java
@@ -1,0 +1,26 @@
+
+/*
+ * @appUtil.java@
+ * Created on 04-Oct-2023
+ *
+ * Copyright (c) 2023 Imspaliwal.
+ * All Rights Reserved.
+ *
+ * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF Imspaliwal
+ * The copyright notice above does not evidence any
+ * actual or intended publication of such source code.
+ *
+ */
+package com.qa.orangehrm.util;
+
+import org.openqa.selenium.WebDriver;
+
+import com.qa.api_common.util.ElementUtil;
+
+public class AppUtil extends ElementUtil {
+
+    public AppUtil(WebDriver driver) {
+        super(driver);
+    }
+
+}
