+++ src/test/java/com/qa/api_common/tests/actions/ElementConcatTests.java
@@ -1,0 +1,5 @@
+package com.qa.api_common.tests.actions;
+
+public class ElementConcatTests {
+
+}
