--- src/test/java/com/qa/orangehrm/tests/firstTests/FirstTest.java
@@ -1,48 +1,0 @@
-
-/*
- * @FirstTest.java@
- * Created on 05-Jul-2023
- *
- * Copyright (c) 2023 Imspaliwal.
- * All Rights Reserved.
- *
- * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF Imspaliwal
- * The copyright notice above does not evidence any
- * actual or intended publication of such source code.
- *
- */
-package com.qa.orangehrm.tests.firstTests;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.testng.annotations.AfterTest;
-import org.testng.annotations.BeforeTest;
-import org.testng.annotations.Test;
-
-public class FirstTest {
-
-    private WebDriver driver;
-
-    @BeforeTest
-    public void setup() {
-        driver = new ChromeDriver();
-    }
-
-    @AfterTest
-    public void tearDown() {
-        if (driver != null) {
-            driver.quit();
-        }
-    }
-
-    @Test
-    public void fisrtTest() {
-
-        // Open Google.com
-        driver.get("https://www.google.com/");
-
-        System.out.println("This is my first test: " + driver.getTitle());
-
-    }
-
-}
