--- src/test/java/com/qa/tests/actions/ElementConcatTests.java
@@ -1,5 +1,0 @@
-package com.qa.tests.actions;
-
-public class ElementConcatTests {
-
-}
